data_source/DriveParsers/requirements.txt
PyPDF2>=3.0.1
"unstructured"
langchain>=0.3.4
langchain_community>=0.3.3
langchain_google_community>=2.0.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googleapiclient
python-dotenv